Refactor tests + alternating projection WIP
