+++ phastphase/retrieval_jax/_alternating_projections.py
@@ -5,6 +5,7 @@ import jax.numpy as jnp
 from jax import lax
 
 from ._loss_funcs import normalized_intensity_loss, sqrt_intensity_loss
+from .alternative_methods._alternating_minimization import Rs, Rm, Pm
 
 
 class _HIOState(NamedTuple):
@@ -57,6 +58,7 @@ def HIO(
     tolerance: float = 1e-3,
     max_iters: int = 100,
     beta: float = 0.9,
+    **kwargs
 ) -> _HIOState:
     """
     Runs the Hybrid Input-Output (HIO) algorithm given an initial seed object `x0`.
@@ -149,6 +151,7 @@ def damped_ER(
     max_iters: int = 100,
     near_field_damping: float = 0.9,
     far_field_damping: float = 0.9,
+    **kwargs
 ) -> _HIOState:
     """
     Perform the Damped Error Reduction (ER) algorithm.
@@ -193,6 +196,132 @@ def damped_ER(
     )
     final_state = lax.while_loop(
         damped_ER_cond_func, damped_ER_step_func, initial_state
+    )
+
+    return final_state
+
+
+def ER(
+    x0: jnp.ndarray,
+    mask: jnp.ndarray,
+    y: jnp.ndarray,
+    tolerance: float = 1e-3,
+    max_iters: int = 100,
+    **kwargs
+) -> _HIOState:
+    """
+    Perform the Error Reduction (ER) algorithm.
+
+    This function implements the Error Reduction (ER) algorithm for phase retrieval.
+    It iteratively updates an initial guess `x0` to minimize the residual error between the
+    measured data `y` and the Fourier transform of the masked input.
+
+    Args:
+        x0 (jnp.ndarray): Initial guess for the solution.
+        mask (jnp.ndarray): Binary mask applied to the input in the near field.
+        y (jnp.ndarray): Measured intensity data in the far field.
+        tolerance (float, optional): Convergence tolerance for the residual error. Defaults to 1e-3.
+        max_iters (int, optional): Maximum number of iterations to perform. Defaults to 100.
+
+    Returns:
+        _HIOState: A dataclass containing the final state of the algorithm, including:
+            - `x` (jnp.ndarray): The reconstructed object.
+            - `iteration` (int): The number of iterations performed.
+            - `residual` (float): The final residual value indicating the difference
+              between the current and target Fourier magnitudes.
+    """
+    return damped_ER(
+        x0=x0,
+        mask=mask,
+        y=y,
+        tolerance=tolerance,
+        max_iters=max_iters,
+        near_field_damping=0.0,
+        far_field_damping=0.0,
+    )
+
+
+def _RAAR_projection_step(
+    state: _HIOState,
+    y: jnp.ndarray,
+    mask: jnp.ndarray,
+    beta: float = 0.9,
+) -> jnp.ndarray:
+    """
+    Perform the Relaxed Averaged Alternating Reflection (RAAR) step.
+
+    Args:
+        state (_HIOState): The current state of the HIO algorithm.
+        y (jnp.ndarray): The target amplitudes.
+        mask (jnp.ndarray): The near-field support mask.
+
+    Returns:
+        jnp.ndarray: The updated estimate of the image.
+    """
+    y_c = jnp.fft.fft2(state.x, norm="ortho")
+    y_c = jnp.sign(y_c) * jnp.sqrt(y)
+    x_new = jnp.fft.ifft2(y_c, norm="ortho")
+    x_new = mask * (x_new) + (1 - mask) * (state.x - beta * x_new)
+
+    x_new = 1 / 2 * beta * (Rs(Rm(state.x, y), mask) + state.x) + (
+        1 - beta
+    ) * Pm(state.x, y)
+
+    residual = normalized_intensity_loss(mask * x_new, y)
+
+    return _HIOState(x=x_new, iteration=state.iteration + 1, residual=residual)
+
+
+def RAAR(
+    x0: jnp.ndarray,
+    mask: jnp.ndarray,
+    y: jnp.ndarray,
+    tolerance: float = 1e-3,
+    max_iters: int = 100,
+    beta: float = 0.9,
+    **kwargs
+) -> _HIOState:
+    """
+    Perform the Relaxed Averaged Alternating Reflections (RAAR) algorithm.
+
+    This function implements the Relaxed Averaged Alternating Reflections (RAAR) algorithm for phase retrieval.
+    It iteratively updates an initial guess `x0` to minimize the residual error between the
+    measured data `y` and the Fourier transform of the masked input.
+
+    Args:
+        x0 (jnp.ndarray): Initial guess for the solution.
+        mask (jnp.ndarray): Binary mask applied to the input in the near field.
+        y (jnp.ndarray): Measured intensity data in the far field.
+        tolerance (float, optional): Convergence tolerance for the residual error. Defaults to 1e-3.
+        max_iters (int, optional): Maximum number of iterations to perform. Defaults to 100.
+        beta (float, optional): Relaxation parameter for the RAAR algorithm. Defaults to 0.9.
+
+    Returns:
+        _HIOState: A dataclass containing the final state of the algorithm, including:
+            - `x` (jnp.ndarray): The reconstructed object.
+            - `iteration` (int): The number of iterations performed.
+            - `residual` (float): The final residual value indicating the difference
+              between the current and target Fourier magnitudes.
+    """
+    RAAR_step_func = jax.tree_util.Partial(
+        _RAAR_projection_step,
+        mask=mask,
+        y=y,
+        beta=beta,
+    )
+    RAAR_cond_func = jax.tree_util.Partial(
+        _HIO_cond_func, tolerance=tolerance, max_iters=max_iters
+    )
+    initial_state = _HIOState(
+        x=x0,
+        iteration=0,
+        residual=jnp.linalg.norm(
+            (jnp.square(jnp.abs(jnp.fft.fft2(mask * x0, norm="ortho"))) - y)
+            / jnp.sqrt(y + 1e-12)
+        ),
+    )
+    final_state = lax.while_loop(
+        RAAR_cond_func, RAAR_step_func, initial_state
     )
 
     return final_state
